Metrics Provider Gateway Groundwork
This commit lays down the basis for implementing the rest of the Metrics code.
The config is able to be read and a different provider is handed back depending on the config.
The config has validation and will throw if invalid parameters are given to it.

For the "enum" code, this relies on a package called `go-enum` to generate the files.
The enums are plugged into the Yaml system and validated at serialization/deserialization time.

It's a bit janky and required a change to the build system, but it is the lowest effort and
safest approach long-term to ensure we're always properly validating inputs. Long term,
Golang is planning to implement enums directly into the language. Until then though, we have
to make do with what's available.


Former-commit-id: 288b12c197a4622e51491991ced0aa9d86be9906
Former-commit-id: 1b5c158aefe04d8286f98821d433a51e8f6031aa