AWS CDK deployment for Metrics Backend
This change adds the basics for deploying the Metrics Backend on to AWS
via Elastic Container Service (ECS) with Fargate (server instances to run Docker containers on).

It's pretty straightforward and simple for the deployment because we're using the example
AWS "ecs-patterns" library that does most of the heavy lifting for us.

This still isn't finished, and I have more work to do with it. But it's the basics.
Tthat's why I'mcommitting this now. I'll probably add more to this branch as I test further.patterns"