Metrics Backend Server Implementation using AWS CDK (#132)
* Base Go CDK Metrics Server Backend

This commit implements a basic Go Cloud Development Kit backend server
that will eventually become the Metrics Server Backend.

Currently, this service is just a fork of the Guestbook example app.
Fortunately, the example app already has a DB setup along with an HTTP
server to process requests.

I have also added a Readme explaining out metrics server. We should take
that document and make it a part of our Docusaurus docs (or link to it).

As a part of this effort, I also made the example app work with our build
process and live alongside our Go codebase. This does mean that all modules
used by this service now live in our shared `go.mod` file; however, that
seemed like an acceptable tradeoff. If you know how to place the dependencies
into a unique `go.mod` file in the `pkg/metricsserverbackend` folder, please
go ahead. I just didn't know how to and it felt non-trivial.jje

Remaining work:
- Replace MySQL with Postgres
- Implement an endpoint that retrieves metrics
- Add a database table that stores the collected data
- Figure out deployment for this (Do we put it in CI? etc)

* AWS CDK deployment for Metrics Backend

This change adds the basics for deploying the Metrics Backend on to AWS
via Elastic Container Service (ECS) with Fargate (server instances to run Docker containers on).

It's pretty straightforward and simple for the deployment because we're using the example
AWS "ecs-patterns" library that does most of the heavy lifting for us.

This still isn't finished, and I have more work to do with it. But it's the basics.
Tthat's why I'mcommitting this now. I'll probably add more to this branch as I test further.patterns"

* Add missing headers

* Working deployed version of metrics server

* Add subdomain for Metrics server and remove Go code

This PR cleans up the Golang code from the original POC.
It also adds Route53 support for subdomains for the metrics server deployment.

* Add license to Package.json

* Replace go modules with original

Former-commit-id: f172065509de12d0dab4a9f07a96c511bba49b23
Former-commit-id: a18c238620713ce8cdca0866b427d21d9cc35eb2