Fix issues with Secure Components margins
This was a real pain in the neck. Basically, the elements weren't properly lining up.
In one case -- the SecureTextArea demo -- the "save" button was overlapping with the actual text input.

This PR cleans up the behavior with the various elements and fixes everything to be properly aligned and functional.

Please take a moment to clone this and test if the demo app works properly for you. Thanks!

P.S. Cameron was here.