Make Demo App Share Sessions
In order for this to work, I had to implement a `connect`-compatible session store.
I forked the store from this NPM module: https://www.npmjs.com/package/connect-sqlite3

It wasn't TypeScript, so I had to convert it to and fix the bugs. It was using some old school ES5
syntax for "classes" (adding functions to the prototype chain). I fixed that and it worked right away.

Now, when you run the demo app and switch between GraphQL and Express, it just works. You only have to sign in once.

Additionally, I also fixed a bug in the Simple Tokenizer that used HTTPS with LocalStack. This was
passing in CI because we disable Chrome Security, but it was failing locally for me. I was able
to verify that it works properly now with this change.


Former-commit-id: 79916c8319f2dc8ffd2f58616fdd97a625abd146
Former-commit-id: 27399c364340b1883a16c1ec1183cfd2db31e9a7