Make demo app sessions shared (#186)
* Make Demo App Share Sessions

In order for this to work, I had to implement a `connect`-compatible session store.
I forked the store from this NPM module: https://www.npmjs.com/package/connect-sqlite3

It wasn't TypeScript, so I had to convert it to and fix the bugs. It was using some old school ES5
syntax for "classes" (adding functions to the prototype chain). I fixed that and it worked right away.

Now, when you run the demo app and switch between GraphQL and Express, it just works. You only have to sign in once.

Additionally, I also fixed a bug in the Simple Tokenizer that used HTTPS with LocalStack. This was
passing in CI because we disable Chrome Security, but it was failing locally for me. I was able
to verify that it works properly now with this change.

* Add info about forking

* Add JSdoc updated comments

* Slight cleanup to appease linter

* Remove log statement

Former-commit-id: fe967cc259388d6c77616282c15a38e3de27a120
Former-commit-id: a462cd95ed95510794b81f3a02d67951680a11f0