Refactor Demo App and simplify code
There is a lot in this change. For the most part though, I just stripped duplicate code and cleaned up the routing.

The app now uses React Router for everything. This was required to fix the weird hash based routing bugs in the app.
There was simple no way to fix them without just refreshing the page twice... It was a reasonable workaround, but it
was still pretty jarring and buggy.

Each page is now nested under a path. There exists:
- `/express`
- `/graphql`
- `/simple`

Also, there is now only one store that is used across the app. There is only one header and sidebar used in the app.
This makes the experience a _lot_ simpler for us to improve on now. We can continue to add functionality into this
app and it will be much cleaner to handle. I tested the functionality and sharing tokens between the backends is fine.

I haven't added the "auto-login" feature yet. I wanted to clean up the sessions and code first. Now it should be
possible without breaking anything to have that implemented.

Fingers crossed to see how badly this breaks CI! I didn't find any bugs in the app from my testing, but please take
a moment to check it also. It's important that this app works well when people try it out.