Add back GraphQL Store
This commit is... a little nuts. It's a clever solution that's type-safe, but that's also what I don't like about it.

We need to have only one store if we want the app to work well. That means we need 2 functions -- one for the Express and GraphQL implementations.

But how do manage that? Do we now have every component in the app aware of the "mode" that the app runs in? That's a leaky abstraction!

So the solution is to create a standard "actions" object that is passed to every component of the app. Components can define which actions they need in their props.

Then we "map" the desired actions (express or graphql) into the object at runtime based on the mode. But, how do we ensure that every time an action is added that it's properly statically typed?

Well, to do that we follow a normal convention for the function. Like this: `saveSsn` and `saveSsnGql`.

To enforce this, I'm using the TypeScript Template literals functionality in the `types.ts` file to enforce that these keys are present on the store implementation. It works beautifully, but the syntax is quite terse.


Former-commit-id: cdbbd4bfd3fa1e883d7afed05f8b6b56b41ad97c
Former-commit-id: 509632817263f7ed0a6327f6f2f2b5424dbaef70