Fix issues with Secure Components margins (#182)
* Fix issues with Secure Components margins

This was a real pain in the neck. Basically, the elements weren't properly lining up.
In one case -- the SecureTextArea demo -- the "save" button was overlapping with the actual text input.

This PR cleans up the behavior with the various elements and fixes everything to be properly aligned and functional.

Please take a moment to clone this and test if the demo app works properly for you. Thanks!

P.S. Cameron was here.

* Fix linter errors

* Code cleanup

Former-commit-id: 17837d3875fac9c19879977520a9f906de35a113
Former-commit-id: 904449d1c01ebea420397f0119064aca839c32fa