Massive amounts of docs changes
I added a bunch of pages, cleaned up a bunch of text based on feedback, and a lot of other tweaks.

@factoidforrest Please give this a solid kick, when you have a chance. I think these docs will help a ton.

There is still more I'd like to do, but this is the core set of changes that I wanted to make. And now there are links to the Live Demo app everywhere :)