Massive amounts of docs changes
I added a bunch of pages, cleaned up a bunch of text based on feedback, and a lot of other tweaks.

@factoidforrest Please give this a solid kick, when you have a chance. I think these docs will help a ton.

There is still more I'd like to do, but this is the core set of changes that I wanted to make. And now there are links to the Live Demo app everywhere :)


Former-commit-id: 43c2b3566a4b6ceb019e85e40b255dd28e92e614
Former-commit-id: 6ea10ef45bc483d7aa31ca60222af08567823614