update Log4Shell logo with new version available under the public domain
There is a seashell made of a scroll (an ancient medium for logging) in the shape of a 'J' to symbolize Java and the 'j' in log4j. The `>` symbolizes the shell prompt that can be accessed by the attacker. The `_` as the cephalopod's pupil symbolizes the shell/terminal cursor. The cephalopod's tentacles spell out `${$$}`, which represents one form of the malicious input string.