Fix some style issues and make ESLint more pleasant
Many of the rules in the ESLint config will no longer apply during dev. Instead, they will automatically be fixed on commit.

In the rare case, you may find yourself needing to manually fix stuff on commit if the `--fix` flag can't do it's job.
For that, I am sorry. But, this should be the "best of both worlds" because you will no longer see red underlines during
dev that are deeply distracting.

In addition, there are some small style tweaks on the front-end as we gear up for launch.


Former-commit-id: 56e553b4e438408bfb1f5e86b3724e06084faa8c
Former-commit-id: c93d64b74190da00557fff16af9398124a83554e